chore: follow-up on mycollection artworks grid comments
